+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-  "cSpell.words": ["Bokeh"]
+  "cSpell.words": [
+    "Bokeh",
+    "Screenshare"
+  ]
 }
