+++ assets/src/features/leaving-page/components/Rating.tsx
