+++ assets/src/features/leaving-page/components/ReviewSentScreen.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from "react";
+
+type ReviewSentScreenProps = {
+
+};
+
+const ReviewSentScreen: FC = () => {
+    return <div>Danke schon</div>;
+}
+
+
+export default ReviewSentScreen;
