+++ assets/src/features/leaving-page/icons/Info.tsx
