+++ assets/src/features/leaving-page/icons/Plus.tsx
