+++ assets/src/features/leaving-page/icons/Send.tsx
