+++ assets/src/features/leaving-page/icons/Star.tsx
